finished

+++ ft_printf.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_printf.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: erikadugar <erikadugar@student.42.fr>      +#+  +:+       +#+        */
+/*   By: etavera- <etavera-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/06 09:12:39 by etavera-          #+#    #+#             */
-/*   Updated: 2023/03/12 22:32:33 by erikadugar       ###   ########.fr       */
+/*   Updated: 2023/03/13 08:32:55 by etavera-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -174,7 +174,7 @@ int	ft_check_specifier(const char spcr, va_list ptr)
 	{
 		if (ptr)
 		{
-				return (putpointer(va_arg(ptr, unsigned long)));
+				return (putpointer(va_arg(ptr, unsigned long long)));
 		}
 		else
 			return (putstr("(nil)"));
@@ -214,20 +214,23 @@ int	ft_printf(const char *fstr, ...)
 	return (j);
 }
 
+// #include <limits.h>
 // int	main(void)
 // {
-// 	int a = 0;
-// 	int *b = &a;
-// 	printf("%p\n",b);
-// 	// ft_printf("%x\n",b);
-// 	ft_printf("%p\n",b);
-// 	ft_printf(" %p ", 0);
+// 	// int a = 0;
+// 	// int *b = &a;
+// 	// printf("%p\n",b);
+// 	// // ft_printf("%x\n",b);
+// 	// ft_printf("%p\n",b);
+// 	// ft_printf(" %p ", 0);
+
 // 	// printf(" %p ", -1);
 // 	// // TEST(3, print(" %p ", 15));
 // 	// // TEST(4, print(" %p ", 16));
 // 	// // TEST(5, print(" %p ", 17));
-// 	// // TEST(6, print(" %p %p ", LONG_MIN, LONG_MAX));
-// 	// // TEST(7, print(" %p %p ", INT_MIN, INT_MAX));
+// 	// ft_printf(" %p %p ", LONG_MIN, LONG_MAX);
+// 	ft_printf(" %p %p \n", (void *)ULONG_MAX, (void *)-ULONG_MAX);
+// 	printf(" %p %p ", (void *)ULONG_MAX, (void *)-ULONG_MAX);
 // 	// // TEST(8, print(" %p %p ", ULONG_MAX, -ULONG_MAX));
 // 	// // TEST(9, print(" %p %p ", 0, 0));
 // }
